+++ __tests__/state/state.js
@@ -1,0 +1,31 @@
+/* jshint -W117 */
+(function () {
+	'use strict';
+
+	const path    = require('path');
+	const assert  = require('yeoman-assert');
+	const helpers = require('yeoman-test');
+
+	const appName = 'Test State';
+
+	jest.setTimeout(require('../../package.json').jestJasmineTimeout);
+
+	describe('generator-cozen-angular:state', () => {
+		beforeAll(() => {
+			return helpers.run(path.join(__dirname, '../../generators/state'))
+				.withPrompts({
+					appName        : appName,
+					isThisYou      : true,
+					authorFirstName: 'Geoffrey',
+					authorLastName : 'Testelin',
+					authorEmail    : 'geoffrey.testelin@gmail.com'
+				});
+		});
+
+		it('create state js file', () => {
+			assert.file([
+				'app/views/main.state.js'
+			]);
+		});
+	});
+})();